+++ kaji/scripts/verify_npm_package.py
@@ -163,7 +163,7 @@ def verify_npm_tarball(tarball: Path, repo: Path) -> None:
 
     if package.get("name") != "@kaji/sdk" or package.get("version") != "0.2.0-beta.2":
         fail("npm package name/version are not the approved beta coordinates")
-    if package.get("license") != "SEE LICENSE IN LICENSE":
+    if package.get("license") != "FSL-1.1-ALv2":
         fail("npm package license metadata is not canonical")
     for target in export_targets(package.get("exports")) + list(
         (package.get("bin") or {}).values()